requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
playwright>=1.40.0
google-generativeai>=0.3.0
supabase>=2.0.3
pydantic>=2.5.0
pydantic[email]>=2.5.0
python-dotenv>=1.0.0
httpx>=0.25.1
bcrypt>=4.1.0
PyJWT>=2.8.0

